+++ packages/effect/src/Iterable.ts
@@ -24,7 +24,7 @@ import type { NoInfer } from "./Types.js"
  * **Note**. `length` is normalized to an integer >= 1.
  *
  * @example
- * import { makeBy } from 'effect/Iterable'
+ * import { makeBy } from "effect/Iterable"
  *
  * assert.deepStrictEqual(Array.from(makeBy(n => n * 2, { length: 5 })), [0, 2, 4, 6, 8])
  *
@@ -56,7 +56,7 @@ export const makeBy = <A>(f: (i: number) => A, options?: {
  * If `end` is omitted, the range will not have an upper bound.
  *
  * @example
- * import { range } from 'effect/Iterable'
+ * import { range } from "effect/Iterable"
  *
  * assert.deepStrictEqual(Array.from(range(1, 3)), [1, 2, 3])
  *
@@ -78,7 +78,7 @@ export const range = (start: number, end?: number): Iterable<number> => {
  * **Note**. `n` is normalized to an integer >= 1.
  *
  * @example
- * import { replicate } from 'effect/Iterable'
+ * import { replicate } from "effect/Iterable"
  *
  * assert.deepStrictEqual(Array.from(replicate("a", 3)), ["a", "a", "a"])
  *
@@ -129,7 +129,7 @@ export const prepend: {
  * Prepends the specified prefix iterable to the beginning of the specified iterable.
  *
  * @example
- * import * as Iterable from "effect/Iterable"
+ * import { Iterable } from "effect"
  *
  * assert.deepStrictEqual(
  *   Array.from(Iterable.prependAll([1, 2], ["a", "b"])),
@@ -806,11 +806,10 @@ export const filterMapWhile: {
  * Retrieves the `Some` values from an `Iterable` of `Option`s.
  *
  * @example
- * import { getSomes } from "effect/Iterable"
- * import { some, none } from "effect/Option"
+ * import { Iterable, Option } from "effect"
  *
  * assert.deepStrictEqual(
- *   Array.from(getSomes([some(1), none(), some(2)])),
+ *   Array.from(Iterable.getSomes([Option.some(1), Option.none(), Option.some(2)])),
  *   [1, 2]
  * )
  *
@@ -823,11 +822,10 @@ export const getSomes: <A>(self: Iterable<Option<A>>) => Iterable<A> = filterMap
  * Retrieves the `Left` values from an `Iterable` of `Either`s.
  *
  * @example
- * import { getLefts } from "effect/Iterable"
- * import { right, left } from "effect/Either"
+ * import { Iterable, Either } from "effect"
  *
  * assert.deepStrictEqual(
- *   Array.from(getLefts([right(1), left("err"), right(2)])),
+ *   Array.from(Iterable.getLefts([Either.right(1), Either.left("err"), Either.right(2)])),
  *   ["err"]
  * )
  *
@@ -840,11 +838,10 @@ export const getLefts = <R, L>(self: Iterable<Either<R, L>>): Iterable<L> => fil
  * Retrieves the `Right` values from an `Iterable` of `Either`s.
  *
  * @example
- * import { getRights } from "effect/Iterable"
- * import { right, left } from "effect/Either"
+ * import { Iterable, Either } from "effect"
  *
  * assert.deepStrictEqual(
- *   Array.from(getRights([right(1), left("err"), right(2)])),
+ *   Array.from(Iterable.getRights([Either.right(1), Either.left("err"), Either.right(2)])),
  *   [1, 2]
  * )
  *