+++ packages/effect/src/Ordering.ts
@@ -35,10 +35,10 @@ export const reverse = (o: Ordering): Ordering => (o === -1 ? 1 : o === 1 ? -1 :
  * @param onGreaterThan - A function that will be called if the `Ordering` parameter is `1`.
  *
  * @example
- * import { match } from "effect/Ordering"
+ * import { Ordering } from "effect"
  * import { constant } from "effect/Function"
  *
- * const toMessage = match({
+ * const toMessage = Ordering.match({
  *   onLessThan: constant('less than'),
  *   onEqual: constant('equal'),
  *   onGreaterThan: constant('greater than')