+++ packages/sql-mssql/src/index.ts
@@ -10,11 +10,6 @@ export * as client from "./Client.js"
 /**
  * @since 1.0.0
  */
-export * as error from "@effect/sql/Error"
-
-/**
- * @since 1.0.0
- */
 export * as migrator from "./Migrator.js"
 
 /**
@@ -26,21 +21,6 @@ export * as parameter from "./Parameter.js"
  * @since 1.0.0
  */
 export * as procedure from "./Procedure.js"
-
-/**
- * @since 1.0.0
- */
-export * as resolver from "@effect/sql/Resolver"
-
-/**
- * @since 1.0.0
- */
-export * as schema from "@effect/sql/Schema"
-
-/**
- * @since 1.0.0
- */
-export * as statement from "@effect/sql/Statement"
 
 export {
   /**