+++ packages/platform-node/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @effect/platform-node
 
+## 0.48.4
+
+### Patch Changes
+
+- Updated dependencies [[`5866c62`](https://github.com/Effect-TS/effect/commit/5866c621d7eb4cc84e4ba972bfdfd219734cd45d)]:
+  - @effect/platform@0.52.3
+  - @effect/platform-node-shared@0.4.18
+
+## 0.48.3
+
+### Patch Changes
+
+- [#2689](https://github.com/Effect-TS/effect/pull/2689) [`e4b82d2`](https://github.com/Effect-TS/effect/commit/e4b82d239bf0974173b5d687b45d1b1899be615f) Thanks [@tim-smart](https://github.com/tim-smart)! - ensure http response middleware is interruptible
+
+- Updated dependencies [[`2e1cdf6`](https://github.com/Effect-TS/effect/commit/2e1cdf67d141281288fffe9a5c10d1379a800513)]:
+  - @effect/platform@0.52.2
+  - effect@3.1.2
+  - @effect/platform-node-shared@0.4.17
+
 ## 0.48.2
 
 ### Patch Changes