+++ packages/typeclass/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @effect/typeclass
 
+## 0.24.11
+
+### Patch Changes
+
+- Updated dependencies [[`2e1cdf6`](https://github.com/Effect-TS/effect/commit/2e1cdf67d141281288fffe9a5c10d1379a800513)]:
+  - effect@3.1.2
+
 ## 0.24.10
 
 ### Patch Changes